+++ changelog.md
@@ -80,3 +80,4 @@
 * Refactor db stocking code
 * Controllers for permissions
 * Edit user passwords on separate page
+* Admin permission required to access admin namespace